requirements.txt
azure-identity
azure-storage-blob
azure-data-tables
azure-common
azure-core
azure-keyvault-secrets